`conversation-activity-filter` - Clean up state
